+++ CHANGELOG.md
@@ -2,14 +2,18 @@
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 # Changelog
 
+- [v1.2.0](#v120)
 - [v1.1.0](#v110)
 - [v1.0.0](#v100)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
 
+### v1.2.0
+  * Use the same strategy to merge `asyncComputed` objects as regular `computed` objects.
+
 ### v1.1.0
 
-  * Handle errors in async computed properties
+  * Handle errors in async computed properties.
 
 ### v1.0.0